// app/utils/enums/future.enum.ts
enum FutureStatus {
  UPCOMING = "UPCOMING",
  ACTIVE = "ACTIVE",
  FINISHED = "FINISHED",
  CANCELLED = "CANCELLED",
  DELETED = "DELETED",
  SETTLEMENT_DONE = "SETTLEMENT_DONE",
  WINNING_OPTION_UPDATED = "WINNING_OPTION_UPDATED",
}

enum FuturePayoutStatus {
  ONGOING = "ONGOING",
  PENDING = "PENDING",
  PROCESSING = "PROCESSING",
  COMPLETED = "COMPLETED",
  PAID = "PAID",
  FAILED = "FAILED",
  LOST = "LOST",
}

export { FutureStatus, FuturePayoutStatus };